requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
fastapi==0.61.2
h11==0.11.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
jsonpath==0.82
loguru==0.5.1
more-itertools==8.6.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pydantic==1.7.2
pyparsing==2.4.7
pytest==6.0.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.2
win32-setctime==1.0.3
xlrd==1.2.0
yagmail==0.11.224
zipp==3.1.0
